refactor(LogFilter): split core
