+++ src/LogFilter/LICENSE
@@ -1,8 +1,7 @@
 /**
  * SPDX-License-Identifier: CC-BY-SA-4.0
- * _addText: '{{Gadget Header|license=CC-BY-SA-4.0}}'
+ * _addText: '{{Gadget Header|url=https://git.qiuwen.net.cn/InterfaceAdmin/QiuwenGadgets/src/branch/master/src/LogFilter|license=CC-BY-SA-4.0}}'
  *
  * @base <https://commons.wikimedia.org/wiki/MediaWiki:Gadget-rightsfilter.js>
  * @source <https://git.qiuwen.net.cn/InterfaceAdmin/QiuwenGadgets/src/branch/master/src/LogFilter>
- * @dependency mediawiki.util
  */
